perf: optimize analytics page — server-side aggregation, smooth charts, lazy recharts
Move all data aggregation from the client to the server query so only
pre-computed chart data and table stats travel over the wire (~25 KB
instead of ~256 KB RSC payload). Implement LTTB downsampling to cap
chart data at 50 points max, producing smooth monotone curves instead
of jagged/noisy lines with 200+ data points. Replace per-point custom
SVG dot renderer (622 <circle> elements) with ReferenceDot only on
peak and promotion/demotion events (~5 dots). Lazy-load recharts via
next/dynamic so the ~360 KB bundle is deferred. Cap rune and champion
stats tables at 10 rows with a 'show more' toggle. Increase rolling
win rate window from 10 to 20 games for smoother trend lines.

Fixes #270
